+++ spring-cloud-stream-binder-jms-activemq/pom.xml
@@ -15,7 +15,7 @@
 
         <dependency>
             <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-stream-binder-jms</artifactId>
+            <artifactId>spring-cloud-stream-binder-jms-common</artifactId>
             <version>${project.version}</version>
         </dependency>
 
@@ -28,13 +28,18 @@
         <!--Test dependencies-->
         <dependency>
             <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-stream-binder-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-stream-binder-jms-activemq-test-support</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-stream-binder-jms-test-support</artifactId>
+            <artifactId>spring-cloud-stream-binder-jms-common-test-support</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
