+++ spring-cloud-stream-binder-jms-common-test-support/src/main/java/org/springframework/cloud/stream/binder/test/integration/EndToEndIntegrationTests.java
