+++ spring-cloud-stream-binder-jms-common/src/main/java/org/springframework/cloud/stream/binder/jms/utils/PartitionAwareJmsSendingMessageHandler.java
