+++ spring-cloud-stream-binder-jms-common/src/test/java/org/springframework/cloud/stream/binder/jms/utils/DestinationNameResolverTest.java
