--- spring-cloud-stream-binder-jms-solace/lib/README.txt
@@ -1,5 +1,0 @@
-The required Solace API and JMS JARs can be downloaded from http://dev.solacesystems.com/downloads/download_jms-api/
-
-At least sol-jms, sol-common and sol-jcsmp are required.
-
-The current version was built against solace API v7.1.2
