--- spring-cloud-stream-binder-jms-solace/pom.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>spring-cloud-stream-binder-jms-parent</artifactId>
-        <groupId>org.springframework.cloud</groupId>
-        <version>1.0.0.BUILD-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>spring-cloud-stream-binder-jms-solace</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-stream-binder-jms</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.solacesystems</groupId>
-            <artifactId>sol-common</artifactId>
-            <version>7.1.2.248</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/lib/sol-common-7.1.2.248.jar</systemPath>
-        </dependency>
-
-        <dependency>
-            <groupId>com.solacesystems</groupId>
-            <artifactId>sol-jcsmp</artifactId>
-            <version>7.1.2.248</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/lib/sol-jcsmp-7.1.2.248.jar</systemPath>
-        </dependency>
-
-        <dependency>
-            <groupId>com.solacesystems</groupId>
-            <artifactId>sol-jms</artifactId>
-            <version>7.1.2.248</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/lib/sol-jms-7.1.2.248.jar</systemPath>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-stream-binder-jms-test-support</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
-
-    </dependencies>
-
-</project>
