--- spring-cloud-stream-binder-jms-solace/src/test/resources/application.yml
@@ -1,14 +1,0 @@
-spring:
-  main:
-    web-environment: false
-
-  # Necessary for org.springframework.cloud.stream.binder.jms.integration testing multiple ApplicationContexts.
-  jmx:
-    enabled: false
-
-solace:
-  host: 192.168.1.3
-  username: admin
-  password: admin
-
-debug: false
