--- spring-cloud-stream-binder-jms-solace/src/test/resources/logback.xml
@@ -1,10 +1,0 @@
-<configuration>
-    <include resource="org/springframework/boot/logging/logback/base.xml" />
-
-    <logger name="com.solacesystems" level="warn"/>
-
-    <root level="info">
-        <appender-ref ref="CONSOLE" />
-    </root>
-
-</configuration>
